Move adding round to seperate action
